+++ builtin/steps/oci/publish/main.go
@@ -11,19 +11,25 @@ import (
 func main() {
 	logger := slog.Default()
 
-	inputs, err := pkg.ParseInputs(os.Args[1:], os.Getenv)
-	if err != nil {
+	if err := run(logger); err != nil {
 		logger.Error("publish", "err", err)
 		os.Exit(1)
+	}
+}
+
+func run(logger *slog.Logger) error {
+	inputs, err := pkg.ParseInputs(os.Args[1:], os.Getenv)
+	if err != nil {
+		return err
 	}
 
 	slog.SetLogLoggerLevel(inputs.LogLevel)
 
-	err = pkg.NewReleaser().Release(context.Background(), inputs.RemoteImageRef, inputs.Common, inputs.PlatformSpecific)
+	imageIndex, err := pkg.NewReleaser().Release(context.Background(), inputs.RemoteImageRef, inputs.Common, inputs.PlatformSpecific)
 	if err != nil {
-		logger.Error("publish", "err", err)
-		os.Exit(1)
+		return err
 	}
 
 	logger.Info("published step", "image", inputs.RemoteImageRef.MajorMinorPatch().Name())
+	return pkg.NewOutputs(inputs.OutputFile).Write(inputs.RemoteImageRef.MajorMinorPatch(), imageIndex)
 }